Refactor by creating package in model package
