+++ data/src/main/java/com/pocs/data/mapper/PostCategoryMapper.kt
@@ -1,6 +1,6 @@
 package com.pocs.data.mapper
 
-import com.pocs.domain.model.PostCategory
+import com.pocs.domain.model.post.PostCategory
 import java.util.*
 
 fun String.toPostCategory() = PostCategory.valueOf(this.uppercase(Locale.getDefault()))
