+++ data/src/main/java/com/pocs/data/mapper/PostMapper.kt
@@ -1,9 +1,7 @@
 package com.pocs.data.mapper
 
 import com.pocs.data.model.PostDto
-import com.pocs.domain.model.Post
-import com.pocs.domain.model.PostCategory
-import java.util.*
+import com.pocs.domain.model.post.Post
 
 fun PostDto.toEntity() = Post(
     id = id,
