+++ data/src/main/java/com/pocs/data/paging/PostPagingSource.kt
@@ -4,7 +4,7 @@ import androidx.paging.PagingSource
 import androidx.paging.PagingState
 import com.pocs.data.api.PostApi
 import com.pocs.data.mapper.toEntity
-import com.pocs.domain.model.Post
+import com.pocs.domain.model.post.Post
 import java.lang.Exception
 import javax.inject.Inject
