--- domain/src/main/java/com/pocs/domain/model/Post.kt
@@ -1,11 +1,0 @@
-package com.pocs.domain.model
-
-data class Post(
-    val id: Int,
-    val title: String,
-    val content: String,
-    val writer: String,
-    val createdAt: String,
-    val updatedAt: String,
-    val category: PostCategory,
-)
