+++ domain/src/main/java/com/pocs/domain/usecase/post/GetAllPostsUseCase.kt
@@ -1,6 +1,6 @@
 package com.pocs.domain.usecase.post
 
-import com.pocs.domain.model.PostCategory
+import com.pocs.domain.model.post.PostCategory
 import com.pocs.domain.repository.PostRepository
 import javax.inject.Inject
