+++ presentation/src/androidTest/java/com/pocs/presentation/PostEditScreenTest.kt
@@ -8,7 +8,7 @@ import androidx.compose.ui.test.junit4.createComposeRule
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
-import com.pocs.domain.model.PostCategory
+import com.pocs.domain.model.post.PostCategory
 import com.pocs.presentation.model.PostEditUiState
 import com.pocs.presentation.view.post.edit.PostEditScreen
 import org.junit.Assert.assertEquals
