+++ presentation/src/main/java/com/pocs/presentation/mapper/PostMapper.kt
@@ -1,6 +1,6 @@
 package com.pocs.presentation.mapper
 
-import com.pocs.domain.model.Post
+import com.pocs.domain.model.post.Post
 import com.pocs.presentation.model.PostItemUiState
 
 fun Post.toUiState() = PostItemUiState(
