+++ presentation/src/main/java/com/pocs/presentation/model/PostCreateUiState.kt
@@ -1,6 +1,6 @@
 package com.pocs.presentation.model
 
-import com.pocs.domain.model.PostCategory
+import com.pocs.domain.model.post.PostCategory
 
 data class PostCreateUiState(
     override val title: String = "",
