+++ presentation/src/main/java/com/pocs/presentation/model/PostDetailUiState.kt
@@ -1,6 +1,6 @@
 package com.pocs.presentation.model
 
-import com.pocs.domain.model.PostCategory
+import com.pocs.domain.model.post.PostCategory
 
 sealed class PostDetailUiState {
