+++ presentation/src/main/java/com/pocs/presentation/view/home/article/ArticleViewModel.kt
@@ -4,7 +4,7 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import androidx.paging.cachedIn
 import androidx.paging.map
-import com.pocs.domain.model.PostCategory
+import com.pocs.domain.model.post.PostCategory
 import com.pocs.domain.usecase.post.GetAllPostsUseCase
 import com.pocs.presentation.mapper.toUiState
 import com.pocs.presentation.model.ArticleUiState
