+++ presentation/src/main/java/com/pocs/presentation/view/post/create/PostCreateViewModel.kt
@@ -4,7 +4,7 @@ import androidx.compose.runtime.MutableState
 import androidx.compose.runtime.State
 import androidx.compose.runtime.mutableStateOf
 import androidx.lifecycle.ViewModel
-import com.pocs.domain.model.PostCategory
+import com.pocs.domain.model.post.PostCategory
 import com.pocs.domain.usecase.post.AddPostUseCase
 import com.pocs.presentation.model.PostCreateUiState
 import dagger.hilt.android.lifecycle.HiltViewModel
