+++ presentation/src/main/java/com/pocs/presentation/view/post/edit/PostEditViewModel.kt
@@ -2,7 +2,7 @@ package com.pocs.presentation.view.post.edit
 
 import androidx.compose.runtime.*
 import androidx.lifecycle.ViewModel
-import com.pocs.domain.model.PostCategory
+import com.pocs.domain.model.post.PostCategory
 import com.pocs.presentation.model.PostEditUiState
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.Dispatchers
