+++ presentation/src/main/java/com/pocs/presentation/view/user/edit/UserEditViewModel.kt
@@ -4,7 +4,7 @@ import androidx.compose.runtime.MutableState
 import androidx.compose.runtime.State
 import androidx.compose.runtime.mutableStateOf
 import androidx.lifecycle.ViewModel
-import com.pocs.domain.model.User
+import com.pocs.domain.model.user.User
 import com.pocs.presentation.model.UserEditUiState
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.delay
