+++ test-library/src/main/java/com/pocs/test_library/mock/Post.kt
@@ -1,7 +1,7 @@
 package com.pocs.test_library.mock
 
-import com.pocs.domain.model.PostCategory
-import com.pocs.domain.model.PostDetail
+import com.pocs.domain.model.post.PostCategory
+import com.pocs.domain.model.post.PostDetail
 
 val mockPostDetail1 = PostDetail(1, "", mockUser1, "", "", "", PostCategory.NOTICE)
 val mockPostDetail2 = PostDetail(2, "hello", mockUser1, "hi", "", "", PostCategory.NOTICE)
