+++ test-library/src/main/java/com/pocs/test_library/mock/User.kt
@@ -1,6 +1,6 @@
 package com.pocs.test_library.mock
 
-import com.pocs.domain.model.User
-import com.pocs.domain.model.UserType
+import com.pocs.domain.model.user.User
+import com.pocs.domain.model.user.UserType
 
 val mockUser1 = User(1, "", "", 1, UserType.MEMBER, "", 30, "", "")
